+++ front/pages/api/w/[wId]/assistant/agent_configurations/index.ts
@@ -17,13 +17,13 @@ import type * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
+import { deprecatedMaybeShadowWriteFirstActionOnAgentConfiguration } from "@app/lib/action_configurations";
 import { trackAssistantCreated } from "@app/lib/amplitude/node";
 import { getAgentUsage } from "@app/lib/api/assistant/agent_usage";
 import {
   createAgentActionConfiguration,
   createAgentConfiguration,
   createAgentGenerationConfiguration,
-  deprecatedMaybeShadowWriteFirstActionOnAgentConfiguration,
   getAgentConfigurations,
   unsafeHardDeleteAgentConfiguration,
 } from "@app/lib/api/assistant/configuration";
